Finished adding required 1.09 features and will now work on adding status information

+++ Assignment-1.09/headers/character.hpp
@@ -9,23 +9,23 @@ protected:
     uint8_t posX;                // X position in the dungeon
     uint8_t posY;                // Y position in the dungeon
     cell_t previousCell;      // The character that was in the cell before the character moved there
-    uint8_t speed;               // Speed of the character
+    int16_t speed;               // Speed of the character
     cell_t cell;                 // The cell type associated with the character
-    u_int16_t HP;
+    int32_t HP;
     string DAM;
     vector<Item> inventory;
 
 public:
     // Constructor
-    Character(uint8_t x, uint8_t y, cell_t previousCell, uint8_t speed, cell_t cell,  u_int16_t HP,  string DAM,  vector<Item> inventory);
+    Character(uint8_t x, uint8_t y, cell_t previousCell, int16_t speed, cell_t cell,  int32_t HP,  string DAM,  vector<Item> inventory);
     
     // Getters
     uint8_t getPosX() const;
     uint8_t getPosY() const;
     cell_t getPreviousCell() const;
-    uint8_t getSpeed() const;
+    int16_t getSpeed() const;
     cell_t getCell() const;
-    u_int16_t getHealth() const;
+    int32_t getHealth() const;
     string getDamage() const;
     vector<Item>& getInventory();
 
@@ -34,9 +34,9 @@ public:
     void setPosX(uint8_t x);
     void setPosY(uint8_t y);
     void setPreviousCell(cell_t previousCell);
-    void setSpeed(uint8_t speed);
+    void setSpeed(int16_t speed);
     void setCell(cell_t cell);
-    void setHealth(u_int16_t hp);
+    void setHealth(int32_t hp);
     void setDamage(const string& dam);
